@@ -4,7 +4,7 @@ version: 1.0.0
 homepage: https://github.com/avbk/the_internet
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   http: ^0.13.0
